Revert section that said implicit parameters must be in a second (or later) group.
It seems that this advice may be outdated. I added an implicit parameter in the first group
then tried to call it with the empty argument list and was shown a warning in IntelliJ 2020.1 as
well as a compiler error in Scala 2.12.12.

```scala
trait LoggerT[F[_]]
def test[F[_]](implicit l: LoggerT[F]): Unit = ()

test()
```

---

Revert "Update 4-concurrency-parallelism.md"

This reverts commit d5ca868366705ae7c5cd9a715578c19348a49a4c.